pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepface>=0.0.95",
    "django>=5.2.5",
    "django-cors-headers>=4.7.0",
    "django-unfold>=0.64.2",
    "djangorestframework>=3.16.1",
    "gunicorn>=23.0.0",
    "python-decouple>=3.8",
    "requests>=2.32.4",
    "tf-keras>=2.19.0",
]
